% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-scales.R
\name{scale_colour_uio}
\alias{scale_colour_uio}
\alias{scale_color_uio}
\alias{scale_fill_uio}
\title{Colour and fill scale constructors for UiO colors}
\usage{
scale_colour_uio(palette = "redblack", discrete = TRUE, reverse = FALSE, ...)

scale_color_uio(palette = "redblack", discrete = TRUE, reverse = FALSE, ...)

scale_fill_uio(palette = "redblack", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in \code{\link{uio_pal}}.}

\item{discrete}{Logical indicating whether colour aesthetic is discrete or not}

\item{reverse}{Logical indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to \code{\link[ggplot2]{discrete_scale}} or
\code{\link[ggplot2]{scale_color_gradientn}}, used respectively when discrete is TRUE or FALSE}
}
\description{
These scale functions help alter ggplot2 colour and fill
to colours in the UiO branding scheme.
See the colour palettes in \code{\link{uio_pal}} for
variations you can choode from.
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(mpg, disp, colour = cyl)) +
   geom_point(size = 5) +
   scale_colour_uio(discrete = FALSE)

ggplot(mtcars, aes(mpg, disp, colour = cyl)) +
   geom_point(size = 5) +
   scale_colour_uio(discrete = FALSE, reverse = TRUE)

ggplot(mtcars, aes(mpg, disp, colour = cyl)) +
   geom_point(size = 5) +
   scale_colour_uio(discrete = FALSE)

ggplot(mtcars, aes(mpg, disp, colour = factor(cyl))) +
   geom_point(size = 5) +
   scale_colour_uio()
}
