% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-themes.R
\name{theme_uio}
\alias{theme_uio}
\alias{theme_uio_dark}
\alias{theme_uio_void}
\alias{theme_uio_dark_void}
\title{ggplot themes for uio}
\usage{
theme_uio(base_size = 10)

theme_uio_dark(base_size = 10)

theme_uio_void(base_size = 10)

theme_uio_dark_void(base_size = 10)
}
\arguments{
\item{base_size}{text size}
}
\description{
A selection of themes that fit with the
UiO branding instructions.
}
\details{
\itemize{
   \item{theme_uio} - base theme for UiO
   \item{theme_uio_dark} - dark/inverted version of base theme
   \item{theme_uio_void} - base theme without axis information or grid
   \item{theme_uio_darl_void} - dark/inverted theme without axis information or grid
}
}
\examples{
\dontrun{
library(ggplot2)

ggplot(mtcars, aes(mpg, disp, colour = cyl)) +
   geom_point(size = 5) +
   theme_uio()

ggplot(mtcars, aes(mpg, disp, colour = cyl)) +
   geom_point(size = 5) +
   theme_uio_dark()

ggplot(mtcars, aes(mpg, disp, colour = cyl)) +
   geom_point(size = 5) +
   theme_uio_void()

ggplot(mtcars, aes(mpg, disp, colour = factor(cyl))) +
   geom_point(size = 5) +
   theme_uio_dark_void()
}
}
