% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uio-palette.R
\name{uio_cols}
\alias{uio_cols}
\title{Function to extract uio colours as hex codes}
\usage{
uio_cols(...)
}
\arguments{
\item{...}{Character names of uio_colors}
}
\description{
Function to extract uio colours as hex codes
}
