% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-themes.R
\name{uio_font}
\alias{uio_font}
\title{UiO theme fonts}
\usage{
uio_font(type = c("body", "title"))
}
\arguments{
\item{type}{body or title font type}
}
\description{
Returns family name of UiO fonts, to
be used in plots. If for instance used the the
ggplot2-theme functions.
}
