% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uio-palette.R
\name{uio_pal}
\alias{uio_pal}
\title{Return function to interpolate a uio color palette}
\usage{
uio_pal(palette = "redblack", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in uio_palettes}

\item{reverse}{Logical indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to \code{\link[grDevices]{colorRampPalette}}}
}
\description{
Return function to interpolate a uio color palette
}
